refactor: 도넛차트 컴포넌트 분리
